requirements.txt
python-telegram-bot==20.7
opencv-python>=4.8.0
mediapipe==0.10.14
numpy>=1.24.0
anthropic>=0.18.1
sqlalchemy==2.0.23
python-dotenv==1.0.0

# BoulderVision - детекция зацепов
roboflow>=1.1.0
supervision>=0.16.0